Change api_client to take a Stream parameter
Rather than a constructor arg.

Also updated the codegen to build a (currently
non-existent) AsyncAPIClientConnection, which
takes a string-based IPv4 address, and port as
the target for the client connection.

Implementation of this code is still TODO. It
will need to properly manage disconnects, etc,
exponential backoff, and such like to be a good
citizen.